+++ lib/cli.dart
@@ -1,14 +1,6 @@
-/**
- * ````dart
- *
- * ````
- */
-
-
-
-
 export 'cli/command.dart';
 export 'cli/flag.dart';
 export 'cli/steps/check.dart';
 export 'cli/steps/install.dart';
 export 'cli/steps/shell.dart';
+export 'response.dart';