+++ CHANGELOG.md
@@ -11,5 +11,19 @@
 - Fixed directory look up of ``Check``.
 - Added ``Check`` unit test.
 
-## 1.0.2
-- Added ``FlowContext.stopWorkflow(Response message)`` to end a workflow on critical errors.
+## 2.0.0
+From now on, the interface will be stable for a longer period. 
+I apologize for the fast change of the last interface, 
+you should be able to depend on Stepflow.
+
+- Added ``runWorkflow()`` and ``runCommand()`` instead of class methods.
+- Added ``Bubble`` & ``Loop``.
+- Added documentation to all source files.
+- Added examples to ``examples/``.
+- Fixed errors with ``Check``.
+- Removed ``AtomicStep``.
+- Replaced configuration and atomization into one unified execution pipeline.
+- Extracted several snippets into their own functions in multiple classes.
+- Added ``critical`` Level to responses.
+- An execution doesn't have to ``return`` a ``Response`` anymore.
+- Added documentation to README.md